+++ lib/src/helpers.dart
@@ -4,11 +4,11 @@ double deviceHeight(context) => MediaQuery.of(context).size.height;
 
 double deviceWidth(context) => MediaQuery.of(context).size.width;
 
-bool compact(context) => deviceWidth(context) < Styles.compact;
+bool isCompact(context) => deviceWidth(context) < Styles.compact;
 
-bool medium(context) => !compact(context) && !expanded(context);
+bool isMedium(context) => !isCompact(context) && !isExpanded(context);
 
-bool expanded(context) => deviceWidth(context) > Styles.expanded;
+bool isExpanded(context) => deviceWidth(context) > Styles.expanded;
 
 BorderRadius borderRadiusAll(double borderRadius) => BorderRadius.all(
       Radius.circular(borderRadius),